+++ sentry_sdk/integrations/aws_lambda.py
@@ -9,7 +9,7 @@ from sentry_sdk.utils import (
     logger,
 )
 from sentry_sdk.integrations import Integration
-from sentry_sdk.integrations._wsgi import _filter_headers
+from sentry_sdk.integrations._wsgi_common import _filter_headers
 
 
 class AwsLambdaIntegration(Integration):
