+++ sentry_sdk/integrations/wsgi.py
@@ -8,9 +8,33 @@ from sentry_sdk.utils import (
     event_from_exception,
 )
 from sentry_sdk._compat import reraise, implements_iterator, text_type
+from sentry_sdk.integrations._wsgi_common import _filter_headers
 
 
-def get_environ(environ):
+class SentryWsgiMiddleware(object):
+    __slots__ = ("app",)
+
+    def __init__(self, app):
+        self.app = app
+
+    def __call__(self, environ, start_response):
+        hub = Hub.current
+        hub.push_scope()
+        with capture_internal_exceptions():
+            with hub.configure_scope() as scope:
+                scope.add_event_processor(_make_wsgi_event_processor(environ))
+
+        try:
+            rv = self.app(environ, start_response)
+        except Exception:
+            einfo = _capture_exception(hub)
+            hub.pop_scope_unsafe()
+            reraise(*einfo)
+
+        return _ScopePoppingResponse(hub, rv)
+
+
+def _get_environ(environ):
     """
     Returns our whitelisted environment variables.
     """
@@ -27,7 +51,7 @@ def get_environ(environ):
 #
 # We need this function because Django does not give us a "pure" http header
 # dict. So we might as well use it for all WSGI integrations.
-def get_headers(environ):
+def _get_headers(environ):
     """
     Returns only proper HTTP headers.
 
@@ -43,104 +67,6 @@ def get_headers(environ):
             yield key.replace("_", "-").title(), value
 
 
-class RequestExtractor(object):
-    def __init__(self, request):
-        self.request = request
-
-    def extract_into_event(self, event):
-        client = Hub.current.client
-        if client is None:
-            return
-
-        content_length = self.content_length()
-        request_info = event.setdefault("request", {})
-        request_info["url"] = self.url()
-
-        if _should_send_default_pii():
-            request_info["cookies"] = dict(self.cookies())
-
-        bodies = client.options["request_bodies"]
-        if (
-            bodies == "never"
-            or (bodies == "small" and content_length > 10 ** 3)
-            or (bodies == "medium" and content_length > 10 ** 4)
-        ):
-            data = AnnotatedValue(
-                "",
-                {"rem": [["!config", "x", 0, content_length]], "len": content_length},
-            )
-        else:
-            parsed_body = self.parsed_body()
-            if parsed_body:
-                data = parsed_body
-            elif self.raw_data():
-                data = AnnotatedValue(
-                    "",
-                    {"rem": [["!raw", "x", 0, content_length]], "len": content_length},
-                )
-            else:
-                return
-
-        request_info["data"] = data
-
-    def content_length(self):
-        try:
-            return int(self.env().get("CONTENT_LENGTH", 0))
-        except ValueError:
-            return 0
-
-    def url(self):
-        raise NotImplementedError()
-
-    def cookies(self):
-        raise NotImplementedError()
-
-    def raw_data(self):
-        raise NotImplementedError()
-
-    def form(self):
-        raise NotImplementedError()
-
-    def parsed_body(self):
-        form = self.form()
-        files = self.files()
-        if form or files:
-            data = dict(form.items())
-            for k, v in files.items():
-                size = self.size_of_file(v)
-                data[k] = AnnotatedValue(
-                    "", {"len": size, "rem": [["!raw", "x", 0, size]]}
-                )
-
-            return data
-
-        return self.json()
-
-    def is_json(self):
-        mt = (self.env().get("CONTENT_TYPE") or "").split(";", 1)[0]
-        return (
-            mt == "application/json"
-            or (mt.startswith("application/"))
-            and mt.endswith("+json")
-        )
-
-    def json(self):
-        try:
-            if self.is_json():
-                raw_data = self.raw_data()
-                if not isinstance(raw_data, text_type):
-                    raw_data = raw_data.decode("utf-8")
-                return json.loads(raw_data)
-        except ValueError:
-            pass
-
-    def files(self):
-        raise NotImplementedError()
-
-    def size_of_file(self, file):
-        raise NotImplementedError()
-
-
 def get_client_ip(environ):
     """
     Naively yank the first IP address in an X-Forwarded-For header
@@ -153,23 +79,6 @@ def get_client_ip(environ):
         return environ["HTTP_X_FORWARDED_FOR"].split(",")[0].strip()
     except (KeyError, IndexError):
         return environ.get("REMOTE_ADDR")
-
-
-def run_wsgi_app(app, environ, start_response):
-    hub = Hub.current
-    hub.push_scope()
-    with capture_internal_exceptions():
-        with hub.configure_scope() as scope:
-            scope.add_event_processor(_make_wsgi_event_processor(environ))
-
-    try:
-        rv = app(environ, start_response)
-    except Exception:
-        einfo = _capture_exception(hub)
-        hub.pop_scope_unsafe()
-        reraise(*einfo)
-
-    return _ScopePoppingResponse(hub, rv)
 
 
 def _capture_exception(hub):
@@ -242,22 +151,11 @@ def _make_wsgi_event_processor(environ):
                 request_info["method"] = environ.get("REQUEST_METHOD")
 
             if "env" not in request_info:
-                request_info["env"] = dict(get_environ(environ))
+                request_info["env"] = dict(_get_environ(environ))
 
             if "headers" not in request_info:
-                request_info["headers"] = _filter_headers(dict(get_headers(environ)))
+                request_info["headers"] = _filter_headers(dict(_get_headers(environ)))
 
         return event
 
     return event_processor
-
-
-def _filter_headers(headers):
-    if _should_send_default_pii():
-        return headers
-
-    return {
-        k: v
-        for k, v in headers.items()
-        if k.lower().replace("_", "-") not in ("set-cookie", "cookie", "authorization")
-    }
